+++ enclone_main/src/lib.rs
@@ -16,6 +16,7 @@ pub mod sec_mem;
 pub mod setup;
 pub mod some_filters;
 pub mod split_orbits;
+pub mod stop;
 pub mod subset;
 pub mod vars;
 pub mod weak_chains;
